Split database modules (#37)
Closes #36
